+++ test/test_ubx_message.cpp
@@ -5,7 +5,7 @@ using namespace ublox::ubx;
 
 TEST(CreateUBXMessage, ACK_ACK)
 {
-    uint16_t payload[2] = {kCLASS_CFG,kCFG_VALDEL};
+    uint8_t payload[2] = {kCLASS_CFG,kCFG_VALDEL};
     UBX_message_t message = create_message(kCLASS_ACK, kACK_ACK, 2, payload);
     ASSERT_EQ(message.message_class, kCLASS_ACK);
     ASSERT_EQ(message.message_id, kACK_ACK);
@@ -16,14 +16,14 @@ TEST(CreateUBXMessage, ACK_ACK)
 
 TEST(CreateUBXMessage, ACK_ACK_Checksum_A)
 {
-    uint16_t payload[2] = {kCLASS_CFG,kCFG_VALDEL};
+    uint8_t payload[2] = {kCLASS_CFG,kCFG_VALDEL};
     UBX_message_t message = create_message(kCLASS_ACK, kACK_ACK, 2, payload);
     ASSERT_EQ(message.get_checksum_a(), 154);
 }
 
 TEST(CreateUBXMessage, ACK_ACK_Checksum_B)
 {
-    uint16_t payload[2] = {kCLASS_CFG,kCFG_VALDEL};
+    uint8_t payload[2] = {kCLASS_CFG,kCFG_VALDEL};
     UBX_message_t message = create_message(kCLASS_ACK, kACK_ACK, 2, payload);
     ASSERT_EQ(message.get_checksum_b(), 195);
 }
